Remove unused code + refactor
