+++ cmd/kubeapps-apis/plugins/pkg/helm/agent/docker_secrets_postrenderer_test.go
