+++ cmd/kubeapps-apis/plugins/pkg/helm/agent/httpConfigForCluster.go
